requirements.txt
fastapi==0.68.0
requests==2.26.0
beautifulsoup4==4.10.0
pydantic==1.10.0
uvicorn[standard]
openai
panda
python-multipart
PyPDF2
python-pptx